fix(app): preserve scroll-away on delayed history
